Move AssetView from assets to mission to eliminate assets → mission/search imports
