+++ assets/urls.py
@@ -10,7 +10,6 @@ from . import views
 urlpatterns = [
     re_path(r'^assets/assettypes/$', views.AssetsTypeView.as_view(), name='asset_types_view'),
     re_path(r'^assets/$', views.AssetsView.as_view(), name='assets_view'),
-    re_path(r'^assets/(?P<asset_id>\d+)/$', views.AssetView.as_view(), name='asset_view'),
     re_path(r'^assets/(?P<asset_id>\d+)/status/$', views.AssetStatusView.as_view(), name='assets_status'),
     re_path(r'^assets/status/values/$', views.assets_status_value_list, name='asset_status_values_list'),
 ]
