+++ assets/views.py
@@ -7,12 +7,7 @@ from django.shortcuts import get_object_or_404, render
 from django.utils.decorators import method_decorator
 from django.views import View
 
-from mission.decorators import mission_asset_get
-from mission.models import AssetCommand
-
-from organization.decorators import asset_is_operator, asset_is_recorder
-from search.models import Search
-from search.view_helpers import check_searches_in_progress
+from organization.decorators import asset_is_recorder
 from .models import AssetType, Asset, AssetStatusValue, AssetStatus
 
 
@@ -22,51 +17,6 @@ def assets_status_value_list(request):
     List all of the asset status values
     """
     return JsonResponse({'values': [v.as_object() for v in AssetStatusValue.objects.all()]})
-
-
-@method_decorator(login_required, name="dispatch")
-@method_decorator(asset_is_operator, name="dispatch")
-class AssetView(View):
-    """
-    View of a specific asset
-    """
-    def as_json(self, request, asset):
-        """
-        Provide the details of an asset
-        """
-        data = {
-            'asset_id': asset.pk,
-            'name': asset.name,
-            'asset_type': asset.asset_type.name,
-            'owner': str(asset.owner),
-            'last_command': AssetCommand.last_command_for_asset_to_json(asset),
-        }
-
-        mission_asset = mission_asset_get(asset)
-        if mission_asset is not None:
-            data['mission_id'] = mission_asset.mission.pk
-            data['mission_name'] = mission_asset.mission.mission_name
-
-            current_search = check_searches_in_progress(mission_asset.mission, asset)
-            if current_search is not None:
-                data['current_search_id'] = current_search.pk
-            queued_search = Search.oldest_queued_for_asset(mission_asset.mission, asset)
-            if queued_search is not None:
-                data['queued_search_id'] = queued_search.pk
-
-        status = AssetStatus.current_for_asset(asset)
-        if status is not None:
-            data['status'] = status.as_object()
-
-        return JsonResponse(data)
-
-    def get(self, request, asset):
-        """
-        View of the specific asset
-        """
-        if "application/json" in request.META.get('HTTP_ACCEPT', ''):
-            return self.as_json(request, asset)
-        return render(request, 'assets/ui.html', {'assetId': asset.pk, 'assetName': asset.name})
 
 
 @method_decorator(login_required, name="dispatch")
