+++ mission/urls.py
@@ -27,4 +27,5 @@ urlpatterns = [
     re_path(r'^$', views.mission_list, name='mission_list'),
     re_path(r'^mission/(?P<mission_id>\d+)/assets/command/set/$', views.AssetCommandSetView.as_view(), name='asset_command_set'),
     re_path(r'^assets/(?P<asset_id>\d+)/command/$', views.AssetCommandView.as_view(), name='assets_command'),
+    re_path(r'^assets/(?P<asset_id>\d+)/$', views.AssetView.as_view(), name='asset_view'),
 ]
